+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Upcoming
 * Animals and sea creatures are now reduced in Winter
+* Bone meal now has 40% and 80% chance to fail to grow plants in autumn and winter respectively
 * Added the `iguanatweaks:doSeasonGrassGrowDeath` gamerule to prevent grass from growing and dying based off seasons
 
 ## 1.1.9